start st4

+++ src/main/java/ru/javacourse/event_notificator/service/NotificationService.java
@@ -82,7 +82,10 @@ public class NotificationService {
         var longs = allUsersNotificationBefore.stream()
                 .map(UserNotificationEntity::getId)
                 .toList();
-        userNotificationRepository.deleteAllById(longs);
+        if (!longs.isEmpty()) {
+            userNotificationRepository.deleteAllById(longs);
+        }
         log.info("deleting {} notifications", longs.size());
+
     }
 }
